refactor TUI code by domain
