+++ cmd/fanout/tui_manual.go
@@ -1,0 +1,259 @@
+package main
+
+import (
+	"bytes"
+	"fmt"
+	"os"
+	"slices"
+	"strings"
+
+	"github.com/butaosuinu/fanout/internal/agent"
+	"github.com/butaosuinu/fanout/internal/cliflags"
+	"github.com/butaosuinu/fanout/internal/exitcode"
+	"github.com/butaosuinu/fanout/internal/hooks"
+	"github.com/butaosuinu/fanout/internal/log"
+	fanoutruntime "github.com/butaosuinu/fanout/internal/runtime"
+	fanouttui "github.com/butaosuinu/fanout/internal/tui"
+)
+
+func newTUILaunchPaneFunc(projectRoot, session, commandName string, hookConfig hooks.Config) fanouttui.LaunchFunc {
+	return func(req fanouttui.LaunchRequest) (string, error) {
+		return launchManualPaneFromTUI(projectRoot, session, commandName, hookConfig, req)
+	}
+}
+
+func launchManualPaneFromTUI(projectRoot, session, commandName string, hookConfig hooks.Config, req fanouttui.LaunchRequest) (string, error) {
+	prompt := normalizeTUIPrompt(req.Prompt)
+	if prompt == "" {
+		return "", fmt.Errorf("prompt is required")
+	}
+	agentNames := normalizeTUIAgents(req.Agents)
+	for _, agentName := range agentNames {
+		if err := agent.ValidateKnown(agentName); err != nil {
+			return "", err
+		}
+		if err := agent.ValidateInstalled(agentName); err != nil {
+			return "", err
+		}
+	}
+	slug, err := normalizeTUISlug(req.Slug)
+	if err != nil {
+		return "", err
+	}
+
+	var stdout, stderr bytes.Buffer
+	launchLogger := log.NewWith(&stdout, &stderr, false)
+	cfg := manualPaneConfigForTUIAgent(agentNames[0])
+	_, recorder, code := loadRunState(cfg, projectRoot, launchLogger)
+	if code != exitcode.OK {
+		return "", bufferedLaunchError(stdout, stderr, "load fanout state")
+	}
+	if recorder != nil {
+		defer func() {
+			_ = recorder.Unlock()
+		}()
+	}
+
+	info := &fanoutruntime.Info{
+		Session:     session,
+		Target:      tuiLaunchTarget(session),
+		ProjectRoot: projectRoot,
+	}
+	createdCount := 0
+	for i, agentName := range agentNames {
+		cfg = manualPaneConfigForTUIAgent(agentName)
+		paneSlug := manualPaneSlugForAgent(slug, agentName, i, agentNames)
+		paneReq := newManualPaneRequest(cfg, projectRoot, recorder.Store, hookConfig, manualPaneOptionsForTUI(prompt, paneSlug, agentName))
+		if createPane(cfg, launchLogger, info, paneReq, recorder, log.Palette{}, commandName) {
+			createdCount++
+			continue
+		}
+		if createdCount > 0 {
+			return partialManualLaunchNotice(createdCount, stderr), nil
+		}
+		return "", bufferedLaunchError(stdout, stderr, "create pane")
+	}
+	return bufferedLaunchNotice(stderr), nil
+}
+
+func partialManualLaunchNotice(createdCount int, stderr bytes.Buffer) string {
+	notice := fmt.Sprintf("created %d new agent pane(s); stopped after a later pane failed", createdCount)
+	if s := strings.TrimSpace(stderr.String()); s != "" {
+		return notice + ": " + compactLaunchError(s)
+	}
+	return notice
+}
+
+func compactLaunchError(s string) string {
+	lines := strings.Split(s, "\n")
+	for _, raw := range slices.Backward(lines) {
+		line := strings.TrimSpace(raw)
+		if strings.HasPrefix(line, "[err ]") {
+			return line
+		}
+	}
+	if len(s) > 180 {
+		return s[:180] + "..."
+	}
+	return s
+}
+
+// bufferedLaunchNotice extracts the tolerated base-refresh skip line, if any,
+// from a successful launch's buffered log so the TUI can show it on success.
+func bufferedLaunchNotice(stderr bytes.Buffer) string {
+	for line := range strings.SplitSeq(stderr.String(), "\n") {
+		if i := strings.Index(line, baseRefreshSkippedNotice); i >= 0 {
+			return strings.TrimSpace(line[i:])
+		}
+	}
+	return ""
+}
+
+func manualPaneConfigForTUIAgent(agentName string) *cliflags.Config {
+	cfg := &cliflags.Config{Agent: agentName}
+	if agentName == "codex" {
+		codexPlanMode := true
+		cfg.CodexPlanMode = &codexPlanMode
+	}
+	return cfg
+}
+
+func manualPaneOptionsForTUI(prompt, slug, agentName string) manualPaneOptions {
+	title := firstPromptLine(prompt)
+	opts := manualPaneOptions{
+		Title:  title,
+		Slug:   slug,
+		Agent:  agentName,
+		Prompt: title,
+	}
+	if strings.Contains(prompt, "\n") {
+		opts.Body = prompt
+	}
+	return opts
+}
+
+func normalizeTUIAgents(raw []string) []string {
+	var agents []string
+	for _, agentName := range raw {
+		agentName = strings.TrimSpace(agentName)
+		if agentName != "" {
+			agents = append(agents, agentName)
+		}
+	}
+	if len(agents) == 0 {
+		return []string{defaultTUIAgent()}
+	}
+	return agents
+}
+
+func manualPaneSlugForAgent(slug, agentName string, index int, agents []string) string {
+	if slug == "" || len(agents) == 1 {
+		return slug
+	}
+	suffix := agentName
+	seen := 0
+	totalForAgent := 0
+	for i, name := range agents {
+		if name != agentName {
+			continue
+		}
+		totalForAgent++
+		if i <= index {
+			seen++
+		}
+	}
+	if totalForAgent > 1 {
+		suffix = fmt.Sprintf("%s-%s", agentName, launchOrdinal(seen))
+	}
+	return strings.TrimRight(slug, "-") + "-" + suffix
+}
+
+func launchOrdinal(n int) string {
+	switch n {
+	case 1:
+		return "one"
+	case 2:
+		return "two"
+	case 3:
+		return "three"
+	default:
+		return fmt.Sprintf("run%d", n)
+	}
+}
+
+func normalizeTUIPrompt(raw string) string {
+	prompt := strings.ReplaceAll(raw, "\r\n", "\n")
+	prompt = strings.ReplaceAll(prompt, "\r", "\n")
+	return strings.TrimSpace(prompt)
+}
+
+func firstPromptLine(prompt string) string {
+	for line := range strings.Lines(prompt) {
+		line = strings.TrimSpace(line)
+		if line != "" {
+			return line
+		}
+	}
+	return strings.TrimSpace(prompt)
+}
+
+func defaultTUIAgent() string {
+	return tuiAgentOrDefault(os.Getenv("FANOUT_AGENT"))
+}
+
+func tuiAgentOrDefault(agentName string) string {
+	if strings.TrimSpace(agentName) == "codex" {
+		return "codex"
+	}
+	return "claude"
+}
+
+func normalizeTUISlug(raw string) (string, error) {
+	slug := strings.TrimSpace(raw)
+	if slug == "" {
+		return "", nil
+	}
+	if !isKebabSlug(slug) {
+		return "", fmt.Errorf("slug must be lowercase kebab-case (alnum+hyphens, starting with alnum), got: %q", slug)
+	}
+	if hasIssueLikeNumericSuffix(slug) {
+		return "", fmt.Errorf("manual slug must not end with an issue-like numeric suffix: %q", slug)
+	}
+	return slug, nil
+}
+
+func isKebabSlug(slug string) bool {
+	for i, r := range slug {
+		switch {
+		case r >= 'a' && r <= 'z', r >= '0' && r <= '9':
+		case r == '-' && i > 0:
+		default:
+			return false
+		}
+	}
+	return slug != ""
+}
+
+func hasIssueLikeNumericSuffix(slug string) bool {
+	i := strings.LastIndex(slug, "-")
+	if i < 0 || i == len(slug)-1 {
+		return false
+	}
+	for _, r := range slug[i+1:] {
+		if r < '0' || r > '9' {
+			return false
+		}
+	}
+	return true
+}
+
+func bufferedLaunchError(stdout, stderr bytes.Buffer, fallback string) error {
+	msg := strings.TrimSpace(stderr.String())
+	if msg == "" {
+		msg = strings.TrimSpace(stdout.String())
+	}
+	if msg == "" {
+		msg = fallback
+	}
+	return fmt.Errorf("%s", msg)
+}
