+++ cmd/fanout/tui_session.go
@@ -1,0 +1,144 @@
+package main
+
+import (
+	"crypto/sha1"
+	"encoding/hex"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/butaosuinu/fanout/internal/exitcode"
+	"github.com/butaosuinu/fanout/internal/log"
+	"github.com/butaosuinu/fanout/internal/tmuxrun"
+)
+
+func tuiLaunchTarget(session string) string {
+	if pane := strings.TrimSpace(os.Getenv("TMUX_PANE")); pane != "" {
+		return pane
+	}
+	return session
+}
+
+func enterTUISession(projectRoot, commandName string, lg *log.Logger) exitcode.Code {
+	session := fanoutTUISessionName(projectRoot)
+	created := false
+	if !tmuxrun.HasSession(session) {
+		if err := tmuxrun.NewSession(session, projectRoot); err != nil {
+			lg.Err("%s", err.Error())
+			return exitcode.Env
+		}
+		created = true
+	}
+
+	pane, running, err := findTUIPane(session)
+	if err != nil {
+		lg.Err("%s", err.Error())
+		return exitcode.Env
+	}
+	if !running {
+		if created {
+			pane, err = firstSessionPane(session)
+		} else {
+			pane, err = tmuxrun.NewWindow(session, tuiPaneTitle, projectRoot)
+		}
+		if err != nil {
+			lg.Err("%s", err.Error())
+			return exitcode.Env
+		}
+		if err := tmuxrun.SendKeys(pane.ID, tuiLaunchCommand(commandName, projectRoot), "Enter"); err != nil {
+			lg.Err("%s", err.Error())
+			return exitcode.Env
+		}
+	}
+	if err := tmuxrun.FocusPane(pane); err != nil {
+		lg.Err("%s", err.Error())
+		return exitcode.Env
+	}
+	if err := tmuxrun.AttachOrSwitch(session); err != nil {
+		lg.Err("%s", err.Error())
+		return exitcode.Env
+	}
+	return exitcode.OK
+}
+
+func markTUIRunning(projectRoot string) func() {
+	paneID := strings.TrimSpace(os.Getenv("TMUX_PANE"))
+	if paneID == "" {
+		return func() {}
+	}
+	_ = tmuxrun.SetPaneProjectRoot(paneID, projectRoot) // Best-effort dashboard keybinding hint.
+	originalTitle, err := tmuxrun.PaneTitle(paneID)
+	if err != nil {
+		originalTitle = "fanout"
+	}
+	_ = tmuxrun.SetPaneTitle(paneID, tuiPaneTitle)
+	return func() {
+		_ = tmuxrun.SetPaneTitle(paneID, originalTitle)
+	}
+}
+
+func findTUIPane(session string) (tmuxrun.PaneInfo, bool, error) {
+	panes, err := tmuxrun.ListPanes(session)
+	if err != nil {
+		return tmuxrun.PaneInfo{}, false, err
+	}
+	for _, pane := range panes {
+		if pane.Title == tuiPaneTitle {
+			return pane, true, nil
+		}
+	}
+	return tmuxrun.PaneInfo{}, false, nil
+}
+
+func firstSessionPane(session string) (tmuxrun.PaneInfo, error) {
+	panes, err := tmuxrun.ListPanes(session)
+	if err != nil {
+		return tmuxrun.PaneInfo{}, err
+	}
+	if len(panes) == 0 {
+		return tmuxrun.PaneInfo{}, fmt.Errorf("tmux session %s has no panes", session)
+	}
+	return panes[0], nil
+}
+
+func tuiProjectRoot() (string, error) {
+	return resolveDisplayProjectRoot()
+}
+
+func fanoutTUISessionName(projectRoot string) string {
+	sum := sha1.Sum([]byte(projectRoot))
+	base := sanitizeSessionPart(filepath.Base(projectRoot))
+	return "fanout-" + base + "-" + hex.EncodeToString(sum[:])[:8]
+}
+
+func sanitizeSessionPart(s string) string {
+	s = strings.ToLower(strings.TrimSpace(s))
+	var b strings.Builder
+	lastDash := false
+	for _, r := range s {
+		allowed := (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9')
+		if allowed {
+			b.WriteRune(r)
+			lastDash = false
+			continue
+		}
+		if !lastDash {
+			b.WriteByte('-')
+			lastDash = true
+		}
+	}
+	out := strings.Trim(b.String(), "-")
+	if out == "" {
+		return "repo"
+	}
+	return out
+}
+
+func tuiLaunchCommand(commandName, projectRoot string) string {
+	exe, err := os.Executable()
+	if err != nil || strings.TrimSpace(exe) == "" {
+		exe = commandName
+	}
+	return "cd " + shellQuote(projectRoot) + " && " + shellQuote(exe)
+}
