+++ cmd/fanout/tui_shell.go
@@ -1,0 +1,122 @@
+package main
+
+import (
+	"crypto/rand"
+	"encoding/hex"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"github.com/butaosuinu/fanout/internal/state"
+	"github.com/butaosuinu/fanout/internal/tmuxrun"
+	fanouttui "github.com/butaosuinu/fanout/internal/tui"
+	"github.com/butaosuinu/fanout/internal/worktree"
+)
+
+func newTUILaunchShellFunc(projectRoot, session string) fanouttui.ShellLaunchFunc {
+	return func(req fanouttui.ShellLaunchRequest) error {
+		return launchShellPaneFromTUI(projectRoot, session, req)
+	}
+}
+
+func launchShellPaneFromTUI(projectRoot, session string, req fanouttui.ShellLaunchRequest) error {
+	rawPath := strings.TrimSpace(req.TargetPath)
+	if rawPath == "" {
+		return fmt.Errorf("terminal path is required")
+	}
+	targetPath, err := filepath.Abs(rawPath)
+	if err != nil {
+		return fmt.Errorf("resolve terminal path: %w", err)
+	}
+	st, statErr := os.Stat(targetPath)
+	if statErr != nil {
+		return fmt.Errorf("terminal path: %w", statErr)
+	} else if !st.IsDir() {
+		return fmt.Errorf("terminal path is not a directory: %s", targetPath)
+	}
+
+	if excludeErr := worktree.EnsureLocalExclude(projectRoot); excludeErr != nil {
+		return fmt.Errorf("prepare local git exclude: %w", excludeErr)
+	}
+
+	recorder, err := state.LockProject(projectRoot)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		_ = recorder.Unlock()
+	}()
+
+	number := nextSyntheticPaneNumber(recorder.Store, manualPaneParentRef)
+	slug := shellPaneSlug(targetPath, req.Root, number)
+	title := shellPaneTitle(targetPath, req.Root)
+	shellKey, err := newShellPaneKey()
+	if err != nil {
+		return err
+	}
+	paneID, err := tmuxrun.SplitPane(tuiLaunchTarget(session), targetPath)
+	if err != nil {
+		return err
+	}
+	if err := tmuxrun.SetPaneShellKey(paneID, shellKey); err != nil {
+		_ = tmuxrun.KillPane(paneID)
+		return err
+	}
+	// Shell pane ergonomics are best-effort; the recorded pane id is enough to
+	// keep the terminal usable when tmux metadata/layout updates fail.
+	_ = tmuxrun.SetPaneTitle(paneID, title)
+	_ = tmuxrun.SetPaneProjectRoot(paneID, projectRoot)
+	_ = tmuxrun.SelectTiled(tuiLaunchTarget(session))
+	if err := recorder.RecordPane(state.Pane{
+		Parent:       manualPaneParentRef,
+		IssueNum:     number,
+		Kind:         state.PaneKindShell,
+		Slug:         slug,
+		PaneID:       paneID,
+		ShellKey:     shellKey,
+		Agent:        state.PaneKindShell,
+		DisplayName:  title,
+		WorktreePath: targetPath,
+		CreatedAt:    time.Now().UTC().Format(time.RFC3339),
+	}); err != nil {
+		_ = tmuxrun.KillPane(paneID)
+		return fmt.Errorf("write fanout state: %w", err)
+	}
+	return nil
+}
+
+func newShellPaneKey() (string, error) {
+	var b [16]byte
+	if _, err := rand.Read(b[:]); err != nil {
+		return "", fmt.Errorf("generate terminal identity: %w", err)
+	}
+	return "shell-" + hex.EncodeToString(b[:]), nil
+}
+
+func shellPaneSlug(targetPath string, root bool, number int) string {
+	base := "root"
+	if !root {
+		base = sanitizeSessionPart(filepath.Base(targetPath))
+	}
+	if base == "" {
+		base = "terminal"
+	}
+	n := number
+	if n < 0 {
+		n = -n
+	}
+	return fmt.Sprintf("terminal-%s-%d", base, n)
+}
+
+func shellPaneTitle(targetPath string, root bool) string {
+	if root {
+		return "root terminal"
+	}
+	base := strings.TrimSpace(filepath.Base(targetPath))
+	if base == "" || base == "." || base == string(filepath.Separator) {
+		base = "worktree"
+	}
+	return "terminal " + base
+}
